project restructure and responsiveness check
